requirements.txt
openai-agents
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0
requests
openai
python-dotenv
jinja2
asyncpg